Latest Commit: Implemented 'double' precision

+++ Raether/src/Aabb.h
@@ -11,11 +11,11 @@ class Aabb {
 public:
 	Interval x, y, z;
 public:
-	Aabb() { } // The default AABB is empty
+	Aabb() {} // The default AABB is empty
 	Aabb(const Interval& x, const Interval& y, const Interval& z) : x(x), y(y), z(z) {
 		PadToMinimum();
 	}
-	Aabb(const glm::vec3 a, const glm::vec3 b) {
+	Aabb(const glm::dvec3 a, const glm::dvec3 b) {
 		x = (a[0] <= b[0]) ? (Interval(a[0], b[0])) : (Interval(b[0], a[0]));
 		y = (a[1] <= b[1]) ? (Interval(a[1], b[1])) : (Interval(b[1], a[1]));
 		z = (a[2] <= b[2]) ? (Interval(a[2], b[2])) : (Interval(b[2], a[2]));
@@ -43,15 +43,15 @@ public:
 	}
 
 	bool Hit(const Ray& ray, Interval& interval) const {
-		const glm::vec3 rayOrigin = ray.Origin;
-		const glm::vec3 rayDirection = ray.Direction;
+		const glm::dvec3 rayOrigin = ray.Origin;
+		const glm::dvec3 rayDirection = ray.Direction;
 
 		for (int axis = 0; axis < 3; axis++) {
 			const Interval& axisInterval = AxisInterval(axis);
-			float rayDirInv = 1.f / rayDirection[axis];
+			double rayDirInv = 1.0 / rayDirection[axis];
 
-			float t0 = (axisInterval.Min - ray.Origin[axis]) * rayDirInv; // axisInterval.Min => x0
-			float t1 = (axisInterval.Max - ray.Origin[axis]) * rayDirInv; // axisInterval.Max => x1
+			double t0 = (axisInterval.Min - ray.Origin[axis]) * rayDirInv; // axisInterval.Min => x0
+			double t1 = (axisInterval.Max - ray.Origin[axis]) * rayDirInv; // axisInterval.Max => x1
 
 			if (t0 < t1) {
 				if (t0 > interval.Min) interval.Min = t0; // interval.Min => t.Start
@@ -73,17 +73,17 @@ public:
 	void PadToMinimum() {
 		// Adjusting the AABB so that no side is narrower than the value Delta
 
-		float Delta = 0.0001f;
+		double Delta = 0.0001;
 		if (x.Size() < Delta) x = x.Expand(Delta);
 		if (y.Size() < Delta) y = y.Expand(Delta);
 		if (z.Size() < Delta) z = z.Expand(Delta);
 	}
 };
 
-inline Aabb operator+(const Aabb& bbox, const glm::vec3& offset) {
+inline Aabb operator+(const Aabb& bbox, const glm::dvec3& offset) {
 	return Aabb(bbox.x + offset.x, bbox.y + offset.y, bbox.z + offset.z);
 }
 
-inline Aabb operator+(const glm::vec3& offset, const Aabb& bbox) {
+inline Aabb operator+(const glm::dvec3& offset, const Aabb& bbox) {
 	return bbox + offset;
 }
